*strn*

+++ libft/ft_strlen.c
@@ -6,11 +6,13 @@
 /*   By: vdescamp <vdescamp@student.42nice.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/18 22:33:44 by vdescamp          #+#    #+#             */
-/*   Updated: 2021/10/20 09:18:53 by vdescamp         ###   ########.fr       */
+/*   Updated: 2021/10/21 00:00:24 by vdescamp         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(const char *s)
+#include <string.h>
+
+size_t	ft_strlen(const char *s)
 {
 	unsigned int	i;
